+++ drizzle/0004_thankful_joshua_kane.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "anime" ADD COLUMN "hidden" boolean DEFAULT false NOT NULL;